Split up C code
