+++ test_audio.rb
@@ -6,17 +6,20 @@ require_relative 'lib/native_audio'
 clip = NativeAudio::Clip.new('knock.wav')
 source = NativeAudio::AudioSource.new(clip)
 
-puts "Playing with Schroeder reverb..."
+puts "Playing with delay + reverb..."
 source.play
 
-# Enable reverb with room-like settings
+# Add a slapback echo
+source.add_delay_tap(time_ms: 200, volume: 0.5)
+
+# Then reverb processes both original and echo
 source.set_reverb(
-  room_size: 0.7,  # 0.0 = small, 1.0 = large hall
-  damping: 0.5,    # High freq decay (0 = bright, 1 = muffled)
-  wet: 0.5,        # Reverb level
-  dry: 0.1         # Original signal level
+  room_size: 0.7,
+  damping: 0.5,
+  wet: 0.5,
+  dry: 0.5
 )
 
-sleep clip.duration + 1.5  # Extra time for reverb tail
+sleep clip.duration + 2.0
 
 puts "Done!"
